fix: fix all search algorithms
