+++ KWNotice/View/SWCentralNotice/SWCentralNoticeViewModel.swift
@@ -33,7 +33,7 @@ final class SWCentralNoticeViewModel: AlertPublishableObject, ObservableObject {
         do {
             notices = try await repository
                 .fetch()
-                .sorted { $0.postedDate > $1.postedDate }
+                .sorted(by: >)
             isAlreadyFetched = true
             state = .none
         } catch {
@@ -49,6 +49,7 @@ final class SWCentralNoticeViewModel: AlertPublishableObject, ObservableObject {
     
     func search(text: String) {
         notices = repository.search(text: text)
+            .sorted(by: >)
     }
     
     func addFavorite(_ notice: SWCentralNotice) {
