+++ KWNoticeKit/Model/Notice/KWHomeNotice.swift
@@ -1,23 +1,27 @@
 //
-//  SWCentralNotice.swift
+//  KWHomeNotice.swift
 //  KWNoticeKit
 //
-//  Created by 김세영 on 2022/07/04.
+//  Created by 김세영 on 2022/07/03.
 //
 
 import Foundation
 
-public struct SWCentralNotice: Identifiable, Decodable {
+public struct KWHomeNotice: Identifiable, Decodable {
     public let id: Int
     public let title: String
+    public let tag: String
     public let postedDate: Date
+    public let modifiedDate: Date
+    public let department: String
     public let url: URL
     public let type: String
     public let crawledTime: Date
     
     enum CodingKeys: String, CodingKey {
-        case id, title, url, type
+        case id, title, tag, department, url, type
         case postedDate = "posted_date"
+        case modifiedDate = "modified_date"
         case crawledTime = "crawled_time"
     }
     
@@ -26,6 +30,8 @@ public struct SWCentralNotice: Identifiable, Decodable {
         
         self.id = try container.decode(Int.self, forKey: .id)
         self.title = try container.decode(String.self, forKey: .title)
+        self.tag = try container.decode(String.self, forKey: .tag)
+        self.department = try container.decode(String.self, forKey: .department)
         self.url = try container.decode(URL.self, forKey: .url)
         self.type = try container.decode(String.self, forKey: .type)
         
@@ -34,13 +40,25 @@ public struct SWCentralNotice: Identifiable, Decodable {
             .yearMonthDateToDate() else {
             throw APIError.decodeError
         }
-        self.postedDate = postedDate
-        
+        guard let modifiedDate = try container
+            .decode(String.self, forKey: .modifiedDate)
+            .yearMonthDateToDate() else {
+            throw APIError.decodeError
+        }
         guard let crawledTime = try container
             .decode(String.self, forKey: .crawledTime)
             .yearMonthDateTimeToDate() else {
             throw APIError.decodeError
         }
+        self.postedDate = postedDate
+        self.modifiedDate = modifiedDate
         self.crawledTime = crawledTime
+    }
+}
+
+extension KWHomeNotice: Comparable {
+    
+    public static func <(lhs: KWHomeNotice, rhs: KWHomeNotice) -> Bool {
+        return lhs.postedDate < rhs.postedDate
     }
 }
